homeView file created with its css and js files
