+++ src/index.html
@@ -4,16 +4,15 @@
     <meta charset="utf-8" />
     <title>Dataverse Chat</title>
     
-
   </head>
   <body>
     
-    <script src="./index.js" type="module"></script>
-
+    <script src="./index.js" type="module" ></script>
     
-          <a href="/" class="redirection">Home</a>
-          <a href="/404" class="redirection">404</a>
-   
+    
+    <a href="/" class="redirection">Home</a>
+    <a href="/404" class="redirection">404</a>
+    
     <div id="root"></div>
   </body>
 </html>
