+++ src/router.js
@@ -1,5 +1,5 @@
-import home from "./viewsComponents/home.js";
-import errorView from "./viewsComponents/errorView.js";
+import home from "./viewsComponents/homeView/home.js";
+import errorView from "./viewsComponents/errorView/errorView.js";
 
 const routes = {
   "/": home,
@@ -7,7 +7,6 @@ const routes = {
 };
 
 function render(template) {
-  console.log("🚀 ~ file: router.js:10 ~ render ~ render:", render)
   const rootElement = document.getElementById("root");
   rootElement.innerHTML = "";
   rootElement.appendChild(template);
@@ -17,9 +16,11 @@ function searchRoute(route) {
   // routes = "/hogar/bebes"
 
   if (routes[route]) {
-    return routes[route]();
+   return routes[route]();
+    
   } else {
-    return routes["/404"]();
+  return routes["/404"]();
+    
   }
 }
