+++ src/viewsComponents/homeView/home.js
@@ -1,16 +1,24 @@
-import data from '../data/dataset.js'
-import { computeStats } from '../lib/dataFunctions.js'
+import data from '../../data/dataset.js'
+import { computeStats } from '../../lib/dataFunctions.js'
+import { createCards } from './createCards.js';
 
 const home = () => {
+   
+    
 
     const link = document.createElement('link');
     link.rel = 'stylesheet';
-    link.href = 'styles.css';
+    link.href = './viewsComponents/homeView/stylesHome.css';
     document.head.appendChild(link);
 
     const container = document.createElement('section')
-    console.log("🚀 ~ file: home.js:6 ~ home ~ container:", container)
+    container.setAttribute("id", "home")
 
+    const buttonToggle = document.createElement("section");
+    buttonToggle.innerHTML = ` 
+    <section class="toggle-btn">
+    <span>&#9776;</span>
+  </section>`;
     
 
     const filtersHtml = `
@@ -76,60 +84,30 @@ const home = () => {
       </section>
     `;
     
-    const ul = document.createElement("ul");
-    ul.setAttribute("class","ulClass")
-    data.forEach(createLi);
-  
-    function createLi(element) {
-      // const li = document.createElement("li");
-      // //Inyectar cada li con los datos de cada elemento
-      const li = `
-      
-      <li class="liClass" data-id="${element.id}" itemscope itemtype="https://schema.org/Person">
-      <section class="card">
-  
-      <section class="front">
-        <dl class="user-name">
-          <figure>
-            <img src="${element.imageUrl}" class="img"/>
-          </figure>
-          <dt></dt><dd itemprop="name">${element.name}</dd>
-          <dt>Canal de televisión: </dt><dd itemprop="channel"> ${element.channel}</dd>
-          <dt>Transmision: </dt><dd itemprop="status">${element.status}</dd>
-          <dt>Audiencia: </dt><dd itemprop="targetAudience">${element.targetAudience}</dd>
-          
-          </section>
-  
-          <section class="back">
-          <dt><b>Género: </b></dt><dd itemprop="genre">${element.genre}</dd>
-          <dt><b>Descripción corta:</b></dt><dd itemprop="shortDescription">${element.shortDescription}</dd>
-          <dt><b>Descripción:</b></dt><dd itemprop="description"> ${element.description}</dd>
-          <dt><b>Hechos importantes:</b> </dt>
-          <dd itemprop="fechaDeCreacionDelPrograma">Fecha de Creacion: ${element.facts.fechaDeCreacionDelPrograma}</dd>
-          <dd itemprop="capituloConMayorRating">Capitulo con mayor rating: ${element.facts.capituloConMayorRating}</dd>
-          <dd itemprop="datoCurioso">Dato curioso: ${element.facts.datoCurioso}</dd>
-  
-          </section>
-        </dl>
-  
-        </section>
-      </li>
-      
-    `;
-  
-      //Cada li inyectarlo en la ul
-      ul.innerHTML += li;
-    }
+    const cards = createCards(data);
 
+    const containerCards = document.createElement("section");
+    containerCards.setAttribute("id", "containerCards");
+    containerCards.appendChild(cards)
+  
     
 
+
+   container.appendChild(buttonToggle);
    container.appendChild(sidebarHtml);
-   container.insertAdjacentElement('beforeend', ul)
+   container.insertAdjacentElement('beforeend', containerCards)
+
+   const script = document.createElement("script");
+    script.src = "./viewsComponents/homeView/mainHome.js"
+    script.type = "module"
+    document.body.appendChild(script)
    
 
     return container
-   
+
     
 };
+
+
 
 export default home;
